requirements.txt
fastapi
uvicorn
python-multipart
pydantic
python-dotenv
PyPDF2
groq